+++ types/CHANGELOG.md
@@ -7,7 +7,9 @@ SPDX-License-Identifier: CC0-1.0
 
 ## [Unreleased](https://github.com/trussed-dev/trussed-rsa-backend/compare/types-v0.1.0...HEAD)
 
--
+- Update dependencies:
+  - `heapless-bytes` v0.5
+  - `trussed-core` v0.2
 
 ## [v0.1.0](https://github.com/trussed-dev/trussed-rsa-backend/releases/tag/types-v0.1.0)